requirements.txt
Flask
google-api-python-client
google-auth
python-dotenv
